# app/ui/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "filmpaw-server",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/performers": {
   "get": {
    "summary": "List Performers",
    "operationId": "list_performers_api_performers_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Q"
      }
     },
     {
      "name": "include_missing",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Include Missing"
      }
     },
     {
      "name": "source_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Source Id"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PerformerListOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/performers/{performer_id}/thumb": {
   "get": {
    "summary": "Get Thumb",
    "operationId": "get_thumb_api_performers__performer_id__thumb_get",
    "parameters": [
     {
      "name": "performer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Performer Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/performers/{performer_id}/aliases": {
   "post": {
    "summary": "Add Alias",
    "operationId": "add_alias_api_performers__performer_id__aliases_post",
    "parameters": [
     {
      "name": "performer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Performer Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AliasIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AliasOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/aliases/{alias_id}": {
   "delete": {
    "summary": "Delete Alias",
    "operationId": "delete_alias_api_aliases__alias_id__delete",
    "parameters": [
     {
      "name": "alias_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Alias Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/performers/{performer_id}": {
   "delete": {
    "summary": "Delete Performer",
    "description": "Single-record delete (UI offers it on missing rows only). D5: group\naliases are intentionally untouched.",
    "operationId": "delete_performer_api_performers__performer_id__delete",
    "parameters": [
     {
      "name": "performer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Performer Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/settings": {
   "get": {
    "summary": "Get Settings",
    "operationId": "get_settings_api_settings_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SettingsOut"
        }
       }
      }
     }
    }
   },
   "put": {
    "summary": "Put Settings",
    "operationId": "put_settings_api_settings_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SettingsIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SettingsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/performers/{performer_id}/open": {
   "post": {
    "summary": "Open Performer",
    "operationId": "open_performer_api_performers__performer_id__open_post",
    "parameters": [
     {
      "name": "performer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Performer Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/local/subdirs": {
   "get": {
    "summary": "Local Subdirs",
    "operationId": "local_subdirs_api_local_subdirs_get",
    "parameters": [
     {
      "name": "path",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Path"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SubdirsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/open-pair": {
   "post": {
    "summary": "Open Pair",
    "operationId": "open_pair_api_open_pair_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OpenPairIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "AliasIn": {
    "properties": {
     "alias": {
      "type": "string",
      "title": "Alias"
     }
    },
    "type": "object",
    "required": [
     "alias"
    ],
    "title": "AliasIn"
   },
   "AliasOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "alias": {
      "type": "string",
      "title": "Alias"
     }
    },
    "type": "object",
    "required": [
     "id",
     "alias"
    ],
    "title": "AliasOut"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "OpenPairIn": {
    "properties": {
     "local_path": {
      "type": "string",
      "title": "Local Path"
     },
     "performer_id": {
      "type": "string",
      "title": "Performer Id"
     }
    },
    "type": "object",
    "required": [
     "local_path",
     "performer_id"
    ],
    "title": "OpenPairIn"
   },
   "PerformerListOut": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/PerformerOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     },
     "source_count": {
      "type": "integer",
      "title": "Source Count"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size",
     "source_count"
    ],
    "title": "PerformerListOut"
   },
   "PerformerOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "source_id": {
      "type": "integer",
      "title": "Source Id"
     },
     "source_label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Label"
     },
     "unc_path": {
      "type": "string",
      "title": "Unc Path"
     },
     "is_missing": {
      "type": "boolean",
      "title": "Is Missing"
     },
     "has_thumb": {
      "type": "boolean",
      "title": "Has Thumb"
     },
     "aliases": {
      "items": {
       "$ref": "#/components/schemas/AliasOut"
      },
      "type": "array",
      "title": "Aliases"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "source_id",
     "source_label",
     "unc_path",
     "is_missing",
     "has_thumb",
     "aliases"
    ],
    "title": "PerformerOut"
   },
...
   "SettingsIn": {
    "properties": {
     "last_local_dir": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Local Dir"
     }
    },
    "type": "object",
    "title": "SettingsIn"
   },
   "SettingsOut": {
    "properties": {
     "last_local_dir": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Local Dir"
     },
     "db_path": {
      "type": "string",
      "title": "Db Path"
     }
    },
    "type": "object",
    "required": [
     "last_local_dir",
     "db_path"
    ],
    "title": "SettingsOut"
   },
...
   "SubdirsOut": {
    "properties": {
     "path": {
      "type": "string",
      "title": "Path"
     },
     "subdirs": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Subdirs"
     }
    },
    "type": "object",
    "required": [
     "path",
     "subdirs"
    ],
    "title": "SubdirsOut"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}